requirements.txt
PySide6
pytest
numpy
Pillow
onnxruntime
transformers
torch
